requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0
mongomock>=4.1.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.1.0
plotly>=5.15.0
pandas>=2.0.0
uuid
datetime
typing 